fff-hoods: Fix resetting interfaces during connection drop
This addresses the following case:
A mesh router has been connected to an uplink router and
downloaded the json. Now the uplink goes down and the mesh router
deletes (!) all interfaces to put up only w2sta. If now the
connection is reestablished, the new json is downloaded and
compared to the still present old one. If the new json does
not differ from the old one, no recreation of wXap and wXmesh
is triggered.
Thus I delete the old file if connection is lost, so comparison
always enters the clause then.

Apply only after the keyV2 patch set!

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>
Tested-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>