Unifi AC: remove bs-partition ro-flag
This patch allows to modify the bs-partition on the AC Mesh.

Without the patch, the bs partition on the AC Mesh is locked
forever. This can be a problem if
a) the bs partition has not been updated correctly during initial flash
b) the sysupgrade mechanism is altered by OpenWRT in order to account for
the flip flop layout

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>
Reviewed-by: Fabian Bläse <fabian@blaese.de>