buildscript: Put binaries in seperate folders per variant.
To make it easier to find the correct binary, this creates
a seperate folder for every variant and copies the binaries
appropriately.

The folder is created with "-p" to suppress errors, if it already
exists. This means the misplaced creation of the "bin"-folder inside
the build function can be ommited.

Signed-off-by: Fabian Bläse <fabian@blaese.de>
Reviewed-by: Robert Langhammer <rlanghammer@web.de>