Replace wpad-mini with hostapd-mini
A wpa_supplicant is not necessary for our firmware,
because it does not connect to encrypted WiFi networks.

Therefore it is possible to use the smaller hostapd-mini
instead of the combined wpad-mini packages.

Some user might use an unsupported setup, where an encrypted
wifi is used for wan uplink. This is not possible anymore.

Signed-off-by: Fabian Bläse <fabian@blaese.de>
Reviewed-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>
Reviewed-by: Robert Langhammer <rlanghammer@web.de>