build_patches: add support for TP-Link CPE210 v3.20
This adds new strings to the support list for the TP-Link CPE210 v3
that are supposed to work with the existing setup.

Without it, the factory image won't be accepted by the vendor UI on
these newer revisions.

This has been merged upstream in commit 4a2380a1e778.

Backport to 19.07 is planned already, so the patch can be removed
again when we move to the next point release.

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>
Acked-by: Fabian Bläse <fabian@blaese.de>